Edit and Delete

+++ app/Http/Controllers/ClassesController.php
@@ -34,10 +34,16 @@ class ClassesController extends Controller
            $data->update();
            return redirect()->route('class.read')->with('success','Class Updated Successfully');
     }
-    public function delete($id){
+    public function delete($id) {
         $data = Classes::find($id);
+    
+        if (!$data) {
+            return redirect()->route('class.read')->with('error', 'Class not found');
+        }
+    
         $data->delete();
-        return redirect()->route('class.read')->with('success','Class Deleted Successfully');
-
+    
+        return redirect()->route('class.read')->with('success', 'Class Deleted Successfully');
     }
+    
 }
